Extract COMPR class
